+++ data/scripts/player/missions/cw_hunter_killer.lua
@@ -1,0 +1,175 @@
+package.path = package.path .. ";data/scripts/lib/?.lua"
+package.path = package.path .. ";data/scripts/?.lua"
+local CosmicVaultFaction = include("cosmicvaultfaction")
+
+include("randomext")
+include("structuredmission")
+
+local MissionUT = include("missionutility")
+local ShipGenerator = include("shipgenerator")
+
+local SectorGenerator = include("SectorGenerator")
+local CosmicWarBridge = include("cosmicwarbridge")
+
+mission._Debug = 0
+mission._Name = "War Contract: Hunter Killer"
+
+mission.data.brief = mission._Name
+mission.data.title = mission._Name
+mission.data.icon = "data/textures/icons/ShipBounty.png"
+mission.data.autoTrackMission = true
+
+local cw_init = initialize
+function initialize(factionIndex)
+    if onServer() and not _restoring then
+        local fIndex = factionIndex
+        local precomputedReward = nil
+        if type(factionIndex) == "table" then
+            fIndex = factionIndex.giver or factionIndex[1]
+            precomputedReward = factionIndex.reward
+        end
+
+        local giverFaction = Faction(fIndex)
+        if not giverFaction then terminate() return end
+
+        local enemyIndex = giverFaction:getValue("enemy_faction") or 0
+        if enemyIndex == 0 then
+            local x, y = Sector():getCoordinates()
+            local pirateLevel = Balancing_GetPirateLevel(x, y)
+            enemyIndex = Galaxy():getPirateFaction(pirateLevel).index
+        end
+
+        mission.data.custom.giverIndex = fIndex
+        mission.data.giver = { factionIndex = fIndex }
+        mission.data.custom.enemyIndex = enemyIndex
+
+        if enemyIndex and enemyIndex > 0 then
+            CosmicVaultFaction.changeRelations(Player().index, enemyIndex, -200000)
+            Player():sendChatMessage(giverFaction.name, 0, "By accepting this contract, you have openly declared war on our enemies."%_T)
+        end
+
+        local x, y = Sector():getCoordinates()
+        local targetX, targetY = MissionUT.getSector(x, y, 2, 10, false, false, false, false, MissionUT.checkSectorInsideBarrier(x, y))
+        if not targetX or not targetY then terminate() return end
+
+        mission.data.location = { x = targetX, y = targetY }
+
+        mission.data.description = {
+            { text = "You accepted a war contract from ${giver}."%_T, arguments = { giver = giverFaction.name } },
+            { text = "An elite enemy strike force is operating in sector (${location.x}:${location.y}). Hunt down the squadron and destroy them."%_T, arguments = { location = mission.data.location } },
+            { text = "Head to sector (${location.x}:${location.y})"%_T, bulletPoint = true, fulfilled = false }
+        }
+
+        local heat = CosmicWarBridge.getFactionWarHeat(fIndex) or 0
+        mission.data.custom.heat = heat
+
+        local baseReward = math.floor(125000 + heat * 150000)
+
+        mission.data.reward = precomputedReward or {
+            credits = baseReward * Balancing_GetSectorRewardFactor(x, y) * ((Faction(mission.data.custom.giverIndex or 0) and Faction(mission.data.custom.giverIndex or 0):getValue("cosmic_trait_cw_mercantile") == 1) and 3 or 1),
+            relations = 10000,
+            paymentMessage = "Contract fulfilled. Payment transferred."%_T
+        }
+
+        cw_init(factionIndex)
+    else
+        cw_init(factionIndex)
+    end
+end
+
+mission.globalPhase.noBossEncountersTargetSector = true
+mission.globalPhase.noPlayerEventsTargetSector = true
+
+mission.phases[1] = {}
+mission.phases[1].showUpdateOnEnd = true
+
+mission.phases[1].onTargetLocationEntered = function(x, y)
+    mission.data.description[3].fulfilled = true
+    if not mission.data.custom.spawned then
+        spawnEvent(x, y)
+        mission.data.custom.spawned = true
+    end
+end
+
+mission.phases[1].triggers = {
+    {
+        condition = function()
+            if onClient() then return false end
+            local _raw_targets = { Sector():getEntitiesByScriptValue("cw_hunter_killer_target") }
+            local targets = {}
+            for _, _t in pairs(_raw_targets) do
+                if _t.type == EntityType.Ship then
+                    table.insert(targets, _t)
+                end
+            end
+            return atTargetLocation() and mission.data.custom.spawned and #targets == 0
+        end,
+        callback = function()
+            reward()
+            accomplish()
+        end
+    }
+}
+
+function spawnEvent(x, y)
+    if onClient() then return end
+
+    local generator = SectorGenerator(x, y)
+    local enemyFaction = Faction(mission.data.custom.enemyIndex)
+    local position = generator:getPositionInSector()
+
+    for i=1, 5 do
+        local ship = ShipGenerator.createMilitaryShip(enemyFaction, position)
+        ship:setValue("cw_hunter_killer_target", true)
+        ship:addScriptOnce("data/scripts/entity/ai/patrol.lua")
+        ship.title = "Elite Strike Force"
+        
+        local ai = ShipAI(ship.index)
+        ai:setAggressive()
+    end
+end
+
+function getBulletin(station)
+    local heat = CosmicWarBridge.getFactionWarHeat(station.factionIndex) or 0
+    if heat < 0.60 then return end
+
+    local baseReward = math.floor(125000 + heat * 150000)
+    local giverFaction = Faction(station.factionIndex)
+    local mult = (giverFaction and giverFaction:getValue("cosmic_trait_cw_mercantile") == 1) and 3 or 1
+    local rewardCredits = baseReward * Balancing_GetSectorRewardFactor(Sector():getCoordinates()) * mult
+    local rewardStruct = {
+        credits = rewardCredits,
+        relations = 10000,
+        paymentMessage = "Contract fulfilled. Payment transferred."%_T
+    }
+
+    return {
+        brief = "War Contract: Hunter Killer"%_T,
+        description = "An elite enemy squadron has been disrupting our operations in a nearby sector. We need you to hunt them down and eliminate them.\n\nWARNING: Accepting this contract is an act of war. You will immediately become hostile to the target faction."%_T,
+        difficulty = "Hard"%_T,
+        reward = "¢${reward}"%_T,
+        script = "data/scripts/player/missions/cw_hunter_killer.lua",
+        icon = "data/textures/icons/ShipBounty.png",
+        formatArguments = { reward = createMonetaryString(rewardCredits) },
+        arguments = { { giver = station.factionIndex, reward = rewardStruct } },
+        msg = "Find that squadron and make sure they don't return. Dismissed."%_T,
+        onAccept = [[
+            local self, player = ...
+            local faction = Faction(self.arguments[1].giver)
+            if faction and player then player:sendChatMessage(faction.name, 0, self.msg) end
+        ]]
+    }
+end
+
+local cw_mission_abandon_original = mission.abandon
+mission.abandon = function()
+    if onServer() then
+        local player = Player()
+        local giverIndex = mission.data.custom.giverIndex
+        if giverIndex and giverIndex > 0 then
+            CosmicVaultFaction.changeRelations(player.index, giverIndex, -25000)
+            player:sendChatMessage(Faction(giverIndex).name, 1, "You abandoned a critical war contract! Our trust in you is broken.")
+        end
+    end
+    if cw_mission_abandon_original then cw_mission_abandon_original() end
+end